# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Тестовое приложение для списка заданий",
        "title": "Список заданий",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/todo": {
            "get": {
                "description": "Получить список элементов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItems"
                ],
                "summary": "Получить список элементов",
                "responses": {}
            },
            "post": {
                "description": "Создать элемент",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItems"
                ],
                "summary": "Создать элемент",
                "parameters": [
                    {
                        "description": "Данные для создания нового элемента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TodoItem"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/todo/{id}": {
            "get": {
                "description": "Получить элемент по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItems"
                ],
                "summary": "Получить элемент по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoItem ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Удалить элемент",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItems"
                ],
                "summary": "Удалить элемент",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoItem ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.TodoItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}